Puffin now just generates depend rules
This is a big commit. It changes puffin to recursively scan for ini
files in a project directory in order to generate a make depend style
auxiliary makefile with rules for building each type of command.

It also adds a default rule list in lib/puffin.ini that is used as the
basis for project ini files.