Added support for hashes of hashes
This is was a big commit that changed the way that the rules adt was
transferred through puffin. Instead of having a flat array with a
single key=value, puffin's adt fields can be defined as hashes or
strings, and hashes can recurse indefinitely. This is a big step
towards being able to parse arbitrary toml.

In addition, this commit adds preliminary support for accessing sub
hashes of data from mustache templates.